+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Do not automatically move map to user location when user is interacting with the map
 
+### Fixed
+- Fixed jumping to user location on every interaction
+
 ## [0.10.0] - 2025-02-28
 
 ### Added