+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - Fixed jumping to user location on every interaction
+- Repeated marker renderings on page load
 
 ## [0.10.0] - 2025-02-28